fix(arcane): P3.6 review -- IslandRootOf walks to UF root (+ island-root test, palette/comment polish)
Finding 1: IslandRootOf previously returned m_uf[i] directly (one-hop parent),
but Island::UpdateSleep uses path-halving so m_uf[i] is a halved intermediate,
not the root. Bodies in the same island at chain depth > 1 returned different
values, silently breaking island-coloring. Fix: walk to the self-pointing root
non-destructively (world is const; m_uf not mutated). Updated doc comment in
PhysicsWorld.hpp to describe the root-walk contract.

Finding 2: Added PhysicsIslandTest.cpp "IslandRootOf returns the same root for
all members of one island" [physics][island]. Uses a 4-box awake stack (checked
while awake so the UF has the contact chain). Confirmed FAILS on old one-hop
impl (root0=2 != root3=4) and PASSES on the fix. rootFar != root0 guards the
separate-island case.

Finding 3: Added fixedRotation capsule comment in PhysicsDebugDraw.cpp so the
intentional angle-not-applied is not mistaken for an omission.

Finding 4: Changed kIslandPalette from `static const` (redundant static inside
anonymous namespace) to `constexpr`, matching the sibling kCol* constants.

Finding 5: Nudged circle B in PhysicsDebugDrawTest from x=60 to x=57 (center
distance 17 < sum of radii 20) so the dynamic-dynamic contact is guaranteed
from step 1 and ForEachContact / magenta contact-line path reliably fires.

+++ Tests/src/PhysicsDebugDrawTest.cpp
@@ -73,11 +73,12 @@ namespace
         circA.density  = Real(1);
         world.AddBody(circA);
 
-        // Dynamic circle B at (60, 90) -- touching / very near A so a contact
-        // pair forms between the two dynamics as they settle.
+        // Dynamic circle B at (57, 90) -- clearly overlapping A (center distance 17 <
+        // sum of radii 20) so a dynamic-dynamic contact is guaranteed from step 1,
+        // making the ForEachContact / magenta contact-line draw path reliably fire.
         BodyDef circB;
         circB.type     = BodyType::Dynamic;
-        circB.position = Vec2(Real(60), Real(90));
+        circB.position = Vec2(Real(57), Real(90));
         circB.shape    = MakeCircle(Real(10));
         circB.density  = Real(1);
         world.AddBody(circB);
